requirements.txt
setuptools
algoliasearch<4.4.0
ook
pybtex
pytest
pylatexenc
pyyaml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
latexcodec
pygments
pre-commit
asyncio
